Extract useCurrentLocation out to utils
