+++ lib/mapUtils.js
@@ -1,6 +1,9 @@
 import * as Analytics from 'expo-firebase-analytics';
+import * as Location from 'expo-location';
+import * as Permissions from 'expo-permissions';
 import convertDistance from 'geolib/es/convertDistance';
 import getDistance from 'geolib/es/getDistance';
+import { useEffect } from 'react';
 import { Alert, Clipboard } from 'react-native';
 import { showLocation } from 'react-native-map-link';
 import {
@@ -334,4 +337,38 @@ export function findDefaultStore(stores) {
     ...deltas,
   };
   return { defaultStore, defaultRegion };
+}
+
+export function useCurrentLocation(setInitialLocation, setlocationErrorMsg) {
+  useEffect(() => {
+    const getCurrentLocation = async () => {
+      try {
+        const { status } = await Permissions.askAsync(Permissions.LOCATION);
+        if (status !== 'granted') {
+          Analytics.setUserProperty('location_permissions', 'denied');
+          throw new Error('Permission to access location was denied');
+        } else {
+          const location = await Location.getCurrentPositionAsync({
+            timeout: 3000,
+          });
+          const currentRegion = {
+            latitude: location.coords.latitude - deltas.latitudeDelta / 3.5,
+            longitude: location.coords.longitude,
+            ...deltas,
+          };
+          setInitialLocation(currentRegion);
+          Analytics.setUserProperty('location_permissions', 'granted');
+        }
+      } catch (err) {
+        Analytics.setUserProperty('location_permissions', 'error');
+        logErrorToSentry({
+          screen: 'MapScreen',
+          function: 'useCurrentLocation',
+          error: err,
+        });
+        setlocationErrorMsg(err.message);
+      }
+    };
+    getCurrentLocation();
+  }, []);
 }
